+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/pipeline/global_nn_recognizer_cvfh.h
@@ -138,9 +138,9 @@ namespace pcl
         bool use_single_categories_;
 
         bool use_cache_;
-        std::map<std::pair<std::string, int>, Eigen::Matrix4f, std::less<std::pair<std::string, int> >, Eigen::aligned_allocator<std::pair<std::pair<
-            std::string, int>, Eigen::Matrix4f> > > poses_cache_;
-        std::map<std::pair<std::string, int>, Eigen::Vector3f> centroids_cache_;
+        std::map<std::pair<std::string, int>, Eigen::Matrix4d, std::less<std::pair<std::string, int> >, Eigen::aligned_allocator<std::pair<std::pair<
+            std::string, int>, Eigen::Matrix4d> > > poses_cache_;
+        std::map<std::pair<std::string, int>, Eigen::Vector3d> centroids_cache_;
 
         std::vector<int> indices_;
 
@@ -188,13 +188,13 @@ namespace pcl
         nearestKSearch (flann::Index<DistT> * index, const flann_model &model, int k, flann::Matrix<int> &indices, flann::Matrix<double> &distances);
 
         void
-        getPose (ModelT & model, int view_id, Eigen::Matrix4f & pose_matrix);
+        getPose (ModelT & model, int view_id, Eigen::Matrix4d & pose_matrix);
 
         bool
-        getRollPose (ModelT & model, int view_id, int d_id, Eigen::Matrix4f & pose_matrix);
+        getRollPose (ModelT & model, int view_id, int d_id, Eigen::Matrix4d & pose_matrix);
 
         void
-        getCentroid (ModelT & model, int view_id, int d_id, Eigen::Vector3f & centroid);
+        getCentroid (ModelT & model, int view_id, int d_id, Eigen::Vector3d & centroid);
 
         void
         getView (ModelT & model, int view_id, PointInTPtr & view);
@@ -202,7 +202,7 @@ namespace pcl
         int NN_;
 
         boost::shared_ptr<std::vector<ModelT> > models_;
-        boost::shared_ptr<std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> > > transforms_;
+        boost::shared_ptr<std::vector<Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d> > > transforms_;
 
         std::vector<double> descriptor_distances_;
 
@@ -337,7 +337,7 @@ namespace pcl
           return models_;
         }
 
-        boost::shared_ptr<std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> > >
+        boost::shared_ptr<std::vector<Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d> > >
         getTransforms ()
         {
           return transforms_;
